dumpling: isolate packed export observability
