+++ dumpling/export/BUILD.bazel
@@ -15,7 +15,9 @@ go_library(
         "metadata.go",
         "metrics.go",
         "packed.go",
+        "packed_observability.go",
         "packed_protocol.go",
+        "packed_protocol_observability.go",
         "prepare.go",
         "retry.go",
         "sql.go",
